requirements.txt
fredapi==0.5.2
numpy>=1.22.4
pandas==2.1.0
scikit-learn==1.3.0
pyyaml==6.0.1
pytest==7.4.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
python-dotenv==1.0.0
PyQt6==6.8.0
imbalanced-learn==0.11.0
joblib>=1.3.0